合并代码

+++ fms-cim-common/src/main/java/org/fms/cim/common/domain/uas/PCalcTaskRelDomain.java
@@ -1,13 +1,13 @@
 package org.fms.cim.common.domain.uas;
 
-import org.fms.cim.common.vo.uas.PCalcTaskRelVO;
-
 import com.riozenc.titanTool.common.reflect.ReflectUtil;
 import com.riozenc.titanTool.mybatis.MybatisEntity;
+import org.fms.cim.common.vo.uas.PCalcTaskRelVO;
+import org.fms.cim.common.vo.uas.PCalcTaskVO;
 
 public class PCalcTaskRelDomain extends PCalcTaskDomain implements MybatisEntity {
     private Long relID;//P_CALC_TPL_DETAIL表ID
-    private Integer tplID;//P_CALC_TPL表的ID
+    private Long tplID;//P_CALC_TPL表的ID
 
     public Long getRelID() {
         return relID;
@@ -17,11 +17,11 @@ public class PCalcTaskRelDomain extends PCalcTaskDomain implements MybatisEntity
         this.relID = relID;
     }
 
-    public Integer getTplID() {
+    public Long getTplID() {
         return tplID;
     }
 
-    public void setTplID(Integer tplID) {
+    public void setTplID(Long tplID) {
         this.tplID = tplID;
     }
