+++ fms-cim-common/src/main/java/org/fms/cim/common/vo/uas/PCalcTplDetailVO.java
@@ -2,21 +2,20 @@
  * 计算方案模板明细
  * Author :
  * Date :
- * Title : org.fms.eis.webapp.vo.PCalcTplDetailVO.java
+ * Title : org.fms.cim.common.vo.uas.PCalcTplDetailVO.java
  **/
 package org.fms.cim.common.vo.uas;
 
-import java.util.Date;
-
-import org.fms.cim.common.domain.uas.PCalcTplDetailDomain;
-
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.riozenc.titanTool.common.reflect.ReflectUtil;
+import org.fms.cim.common.domain.uas.PCalcTplDetailDomain;
+
+import java.util.Date;
 
 public class PCalcTplDetailVO extends ManagerParamVO {
 
     private Long id;    //标识
-    private Integer tplId;    //模板标识(P_CALC_TPL)
+    private Long tplId;    //模板标识(P_CALC_TPL)
     private Long taskId;    //任务标识(P_CALC_TASK)
     private Long creatorId;    //创建者
 
@@ -33,11 +32,11 @@ public class PCalcTplDetailVO extends ManagerParamVO {
         this.id = id;
     }
 
-    public Integer getTplId() {
+    public Long getTplId() {
         return tplId;
     }
 
-    public void setTplId(Integer tplId) {
+    public void setTplId(Long tplId) {
         this.tplId = tplId;
     }
